From sauerkrause/vbjin-ovr@3827452: Use SetWindowLongPtr etc.
Use the proper Win64 typedefs that were introduced to the Windows SDK
in 2003.

Also note I'm NOT basing on saurkrause's repo, as the next commit in
that branch is not something I'd like to do.  It wraps GWLP/SWLP in a
macro depending on _WIN64.  This is bad Win32 programmer mojo!  The
2003-era recommendations for Win64 are to use the new types and let
the platform headers give you the right linkage when building for
32-bit.  The change in that branch only makes sense when using a
very obsolete and irrelevant Windows SDK, as in pre-2003.